Refactor get visit(s) calls
