+++ server/data/visitSchedulerApiClient.ts
@@ -1,51 +1,13 @@
 import { URLSearchParams } from 'url'
 import RestClient from './restClient'
-import { Visit, VisitSession, SessionCapacity, PageVisitDto, SessionSchedule } from './orchestrationApiTypes'
+import { VisitSession, SessionCapacity, SessionSchedule } from './orchestrationApiTypes'
 import config, { ApiConfig } from '../config'
 
 export default class VisitSchedulerApiClient {
   private restClient: RestClient
 
-  private visitType = 'SOCIAL'
-
   constructor(token: string) {
     this.restClient = new RestClient('visitSchedulerApiClient', config.apis.visitScheduler as ApiConfig, token)
-  }
-
-  // Workaround for pagination, mentioned in comments - VB-1760
-  private page = '0'
-
-  private size = '1000'
-
-  async getVisit(reference: string): Promise<Visit> {
-    return this.restClient.get({ path: `/visits/${reference}` })
-  }
-
-  async getUpcomingVisits(offenderNo: string, visitStatus: Visit['visitStatus'][]): Promise<PageVisitDto> {
-    return this.restClient.get({
-      path: '/visits/search',
-      query: new URLSearchParams({
-        prisonerId: offenderNo,
-        startDateTime: new Date().toISOString(),
-        visitStatus: visitStatus.join(','),
-        page: this.page,
-        size: this.size,
-      }).toString(),
-    })
-  }
-
-  async getVisitsByDate(dateString: string, prisonId: string): Promise<PageVisitDto> {
-    return this.restClient.get({
-      path: '/visits/search',
-      query: new URLSearchParams({
-        prisonId,
-        startDateTime: `${dateString}T00:00:00`,
-        endDateTime: `${dateString}T23:59:59`,
-        visitStatus: 'BOOKED',
-        page: this.page,
-        size: this.size,
-      }).toString(),
-    })
   }
 
   async getVisitSessions(offenderNo: string, prisonId: string): Promise<VisitSession[]> {
