+++ server/routes/prisoner.test.ts
@@ -10,7 +10,7 @@ import {
   createMockAuditService,
   createMockPrisonerProfileService,
   createMockPrisonerSearchService,
-  createMockVisitSessionsService,
+  createMockVisitService,
 } from '../services/testutils/mocks'
 
 let app: Express
@@ -20,7 +20,7 @@ let flashData: FlashData
 const auditService = createMockAuditService()
 const prisonerProfileService = createMockPrisonerProfileService()
 const prisonerSearchService = createMockPrisonerSearchService()
-const visitSessionsService = createMockVisitSessionsService()
+const visitService = createMockVisitService()
 
 const prisonId = 'HEI'
 
@@ -41,7 +41,7 @@ beforeEach(() => {
   visitSessionData = {}
 
   app = appWithAllRoutes({
-    services: { auditService, prisonerProfileService, prisonerSearchService, visitSessionsService },
+    services: { auditService, prisonerProfileService, prisonerSearchService, visitService },
     sessionData: {
       visitSessionData,
     } as SessionData,
@@ -485,7 +485,7 @@ describe('GET /prisoner/A1234BC/visits', () => {
     ]
 
     prisonerSearchService.getPrisoner.mockResolvedValue(prisoner)
-    visitSessionsService.getUpcomingVisits.mockResolvedValue(visitInfo)
+    visitService.getUpcomingVisits.mockResolvedValue(visitInfo)
 
     return request(app)
       .get('/prisoner/A1234BC/visits')
@@ -530,7 +530,7 @@ describe('GET /prisoner/A1234BC/visits', () => {
     ]
 
     prisonerSearchService.getPrisoner.mockResolvedValue(prisoner)
-    visitSessionsService.getUpcomingVisits.mockResolvedValue(visitInfo)
+    visitService.getUpcomingVisits.mockResolvedValue(visitInfo)
 
     return request(app)
       .get('/prisoner/A1234BC/visits?search=A1234BC')
@@ -555,7 +555,7 @@ describe('GET /prisoner/A1234BC/visits', () => {
 
   it('should show message and back-to-start button if prisoner has no upcoming visits', () => {
     prisonerSearchService.getPrisoner.mockResolvedValue(prisoner)
-    visitSessionsService.getUpcomingVisits.mockResolvedValue([])
+    visitService.getUpcomingVisits.mockResolvedValue([])
 
     return request(app)
       .get('/prisoner/A1234BC/visits')
