+++ server/routes/visits.test.ts
@@ -10,6 +10,7 @@ import { getParsedDateFromQueryString } from '../utils/utils'
 import {
   createMockAuditService,
   createMockPrisonerSearchService,
+  createMockVisitService,
   createMockVisitSessionsService,
 } from '../services/testutils/mocks'
 
@@ -19,6 +20,7 @@ let flashData: FlashData
 
 const auditService = createMockAuditService()
 const prisonerSearchService = createMockPrisonerSearchService()
+const visitService = createMockVisitService()
 const visitSessionsService = createMockVisitSessionsService()
 
 beforeEach(() => {
@@ -26,7 +28,7 @@ beforeEach(() => {
   flashProvider.mockImplementation((key: keyof FlashData) => {
     return flashData[key]
   })
-  app = appWithAllRoutes({ services: { auditService, prisonerSearchService, visitSessionsService } })
+  app = appWithAllRoutes({ services: { auditService, prisonerSearchService, visitService, visitSessionsService } })
 })
 
 afterEach(() => {
@@ -210,7 +212,7 @@ describe('GET /visits', () => {
 
   it('should render visit slot summary page with prisoner list, slot details and menu for choosing other slots, with the first slot chosen by default ', () => {
     prisonerSearchService.getPrisonersByPrisonerNumbers.mockResolvedValue(prisoners)
-    visitSessionsService.getVisitsByDate.mockResolvedValue(visits)
+    visitService.getVisitsByDate.mockResolvedValue(visits)
 
     return request(app)
       .get('/visits')
@@ -243,10 +245,10 @@ describe('GET /visits', () => {
   it('should not show capacity if session capacity not available', () => {
     prisonerSearchService.getPrisonersByPrisonerNumbers.mockResolvedValue(prisoners)
     visitSessionsService.getVisitSessionCapacity.mockResolvedValue(null)
-    visitSessionsService.getVisitsByDate.mockResolvedValue(visits)
+    visitService.getVisitsByDate.mockResolvedValue(visits)
 
     app = appWithAllRoutes({
-      services: { auditService, prisonerSearchService, visitSessionsService },
+      services: { auditService, prisonerSearchService, visitService, visitSessionsService },
       sessionData: { selectedEstablishment: { prisonId: 'XYZ', prisonName: 'XYZ' } } as SessionData,
     })
 
@@ -263,7 +265,7 @@ describe('GET /visits', () => {
 
   it('should render visit slot summary page with prisoner list, slot details and menu for choosing other slots, with the chosen slot shown for OPEN', () => {
     prisonerSearchService.getPrisonersByPrisonerNumbers.mockResolvedValue(prisoners)
-    visitSessionsService.getVisitsByDate.mockResolvedValue(visits)
+    visitService.getVisitsByDate.mockResolvedValue(visits)
 
     return request(app)
       .get('/visits?selectedDate=2022-05-23&time=10am%20to%2011am&type=OPEN')
@@ -295,7 +297,7 @@ describe('GET /visits', () => {
 
   it('should render visit slot summary page with prisoner list, slot details and menu for choosing other slots, with the chosen slot shown for UNKNOWN', () => {
     prisonerSearchService.getPrisonersByPrisonerNumbers.mockResolvedValue(prisoners)
-    visitSessionsService.getVisitsByDate.mockResolvedValue(visits)
+    visitService.getVisitsByDate.mockResolvedValue(visits)
 
     return request(app)
       .get('/visits?selectedDate=2022-05-23&time=10am+to+11am&type=UNKNOWN')
@@ -327,7 +329,7 @@ describe('GET /visits', () => {
 
   it('should render visit slot summary page with prisoner list, slot details and menu for choosing other slots, with the chosen slot shown, bad date defaults to today', () => {
     prisonerSearchService.getPrisonersByPrisonerNumbers.mockResolvedValue(prisoners)
-    visitSessionsService.getVisitsByDate.mockResolvedValue(visits)
+    visitService.getVisitsByDate.mockResolvedValue(visits)
 
     return request(app)
       .get('/visits?selectedDate=2022-77-23&time=10am%20to%2011am&type=OPEN')
@@ -365,7 +367,7 @@ describe('GET /visits', () => {
       next: 0,
       previous: 0,
     })
-    visitSessionsService.getVisitsByDate.mockResolvedValue({
+    visitService.getVisitsByDate.mockResolvedValue({
       extendedVisitsInfo: [],
       slots: {
         openSlots: [],
@@ -396,7 +398,7 @@ describe('GET /visits', () => {
 
   it('should render no slots message when bad slot', () => {
     prisonerSearchService.getPrisonersByPrisonerNumbers.mockResolvedValue(prisoners)
-    visitSessionsService.getVisitsByDate.mockResolvedValue(visits)
+    visitService.getVisitsByDate.mockResolvedValue(visits)
 
     return request(app)
       .get('/visits?selectedDate=2022-05-23&time=11am%20to%2011am&type=OPEN')
