+++ server/services/testutils/mocks.ts
@@ -37,4 +37,4 @@ export const createMockUserService = () => new UserService(null, null) as jest.M
 export const createMockVisitService = () => new VisitService(null, null, null, null) as jest.Mocked<VisitService>
 
 export const createMockVisitSessionsService = () =>
-  new VisitSessionsService(null, null, null, null) as jest.Mocked<VisitSessionsService>
+  new VisitSessionsService(null, null, null) as jest.Mocked<VisitSessionsService>
