+++ server/services/visitService.test.ts
@@ -1,5 +1,12 @@
-import { VisitSessionData, VisitorListItem } from '../@types/bapv'
-import { OutcomeDto, Visit, VisitHistoryDetails } from '../data/orchestrationApiTypes'
+import { NotFound } from 'http-errors'
+import {
+  ExtendedVisitInformation,
+  VisitInformation,
+  VisitSessionData,
+  VisitorListItem,
+  VisitsPageSlot,
+} from '../@types/bapv'
+import { OutcomeDto, PageVisitDto, Visit, VisitHistoryDetails } from '../data/orchestrationApiTypes'
 import TestData from '../routes/testutils/testData'
 import VisitService from './visitService'
 import {
@@ -8,6 +15,7 @@ import {
   createMockPrisonerContactRegistryApiClient,
 } from '../data/testutils/mocks'
 import { createMockAdditionalSupportService } from './testutils/mocks'
+import { Address, AddressUsage, Contact, Restriction } from '../data/prisonerContactRegistryApiTypes'
 
 const token = 'some token'
 
@@ -366,7 +374,41 @@ describe('Visit service', () => {
     })
   })
 
-  describe('Get Visit', () => {
+  describe('Get Visit(s)', () => {
+    const visit = TestData.visit()
+
+    describe('getVisit', () => {
+      it('should return VisitInformation given a visit reference and matching prisonId', async () => {
+        orchestrationApiClient.getVisit.mockResolvedValue(visit)
+        const result = await visitService.getVisit({ username: 'user', reference: 'ab-cd-ef-gh', prisonId })
+
+        expect(orchestrationApiClient.getVisit).toHaveBeenCalledTimes(1)
+        expect(result).toEqual(<VisitInformation>{
+          reference: 'ab-cd-ef-gh',
+          prisonNumber: 'A1234BC',
+          prisonerName: '',
+          mainContact: 'Jeanette Smith',
+          visitDate: '14 January 2022',
+          visitTime: '10am to 11am',
+          visitStatus: 'BOOKED',
+        })
+      })
+
+      it('should throw a 404 if the visit is not at the given prisonId', async () => {
+        orchestrationApiClient.getVisit.mockResolvedValue(visit)
+
+        await expect(async () => {
+          await visitService.getVisit({
+            username: 'user',
+            reference: 'ab-cd-ef-gh',
+            prisonId: 'BLI',
+          })
+
+          expect(orchestrationApiClient.getVisit).toHaveBeenCalledTimes(1)
+        }).rejects.toBeInstanceOf(NotFound)
+      })
+    })
+
     describe('getFullVisitDetails', () => {
       const visitHistoryDetails = TestData.visitHistoryDetails({
         visit: TestData.visit({
@@ -430,6 +472,302 @@ describe('Visit service', () => {
         expect(additionalSupportService.getAvailableSupportOptions).toHaveBeenCalledTimes(1)
         expect(orchestrationApiClient.getVisitHistory).toHaveBeenCalledTimes(1)
         expect(result).toEqual(expectedResult)
+      })
+    })
+
+    describe('getUpcomingVisits', () => {
+      it('should return an array of upcoming VisitInformation for an offender', async () => {
+        const visits: Visit[] = [visit]
+
+        orchestrationApiClient.getUpcomingVisits.mockResolvedValue({ content: visits })
+        const result = await visitService.getUpcomingVisits({
+          username: 'user',
+          offenderNo: 'A1234BC',
+          visitStatus: ['BOOKED'],
+        })
+
+        expect(orchestrationApiClient.getUpcomingVisits).toHaveBeenCalledTimes(1)
+        expect(orchestrationApiClient.getUpcomingVisits).toHaveBeenCalledWith('A1234BC', ['BOOKED'])
+        expect(result).toEqual(<VisitInformation[]>[
+          {
+            reference: 'ab-cd-ef-gh',
+            prisonNumber: 'A1234BC',
+            prisonerName: '',
+            mainContact: 'Jeanette Smith',
+            visitDate: '14 January 2022',
+            visitTime: '10am to 11am',
+            visitStatus: 'BOOKED',
+          },
+        ])
+      })
+
+      it('should return an empty array for an offender with no upcoming visits', async () => {
+        const visits: Visit[] = []
+
+        orchestrationApiClient.getUpcomingVisits.mockResolvedValue({ content: visits })
+        const result = await visitService.getUpcomingVisits({
+          username: 'user',
+          offenderNo: 'A1234BC',
+          visitStatus: ['BOOKED'],
+        })
+
+        expect(orchestrationApiClient.getUpcomingVisits).toHaveBeenCalledTimes(1)
+        expect(orchestrationApiClient.getUpcomingVisits).toHaveBeenCalledWith('A1234BC', ['BOOKED'])
+        expect(result).toEqual([])
+      })
+    })
+
+    describe('getVisitsByDate', () => {
+      it('should return empty data if no visit sessions on chosen date', async () => {
+        orchestrationApiClient.getVisitsByDate.mockResolvedValue({ content: [] })
+        prisonerContactRegistryApiClient.getPrisonerSocialContacts.mockResolvedValue([])
+        const results = await visitService.getVisitsByDate({
+          username: 'user',
+          dateString: '2022-01-01',
+          prisonId,
+        })
+
+        expect(orchestrationApiClient.getVisitsByDate).toHaveBeenCalledTimes(1)
+        expect(results).toEqual({
+          extendedVisitsInfo: [],
+          slots: {
+            closedSlots: [],
+            firstSlotTime: undefined,
+            openSlots: [],
+            unknownSlots: [],
+          },
+        })
+      })
+
+      it('should return visit and prisoner data when session exists', async () => {
+        const emptyAddresses: Address[] = []
+        const emptyUsages: AddressUsage[] = []
+        const emptyRestrictions: Restriction[] = []
+
+        const pagedVisit: PageVisitDto = {
+          totalPages: 1,
+          totalElements: 1,
+          size: 1,
+          content: [
+            {
+              applicationReference: 'aaa-bbb-ccc',
+              reference: 'ob-cw-lx-na',
+              prisonerId: 'A8709DY',
+              prisonId: 'HEI',
+              visitRoom: 'VISITS-VISITS-H1_6LV',
+              visitType: 'SOCIAL',
+              visitStatus: 'BOOKED',
+              outcomeStatus: 'NOT_RECORDED',
+              visitRestriction: 'OPEN',
+              startTimestamp: '2022-05-23T09:00:00',
+              endTimestamp: '2022-05-23T09:29:00',
+              visitNotes: [],
+              visitContact: {
+                name: 'UNKNOWN',
+                telephone: 'UNKNOWN',
+              },
+              visitors: [
+                {
+                  nomisPersonId: 4729510,
+                },
+              ],
+              visitorSupport: [],
+              createdBy: 'user1',
+              createdTimestamp: '2022-05-23T10:09:56.636334',
+              modifiedTimestamp: '2022-05-23T10:09:56.64691',
+            },
+            {
+              applicationReference: 'aaa-bbb-ccc',
+              reference: 'lb-co-bn-oe',
+              prisonerId: 'A8709DY',
+              prisonId: 'HEI',
+              visitRoom: 'daily test room',
+              visitType: 'SOCIAL',
+              visitStatus: 'BOOKED',
+              outcomeStatus: 'ADMINISTRATIVE_ERROR',
+              visitRestriction: 'OPEN',
+              startTimestamp: '2022-05-23T10:00:00',
+              endTimestamp: '2022-05-23T11:00:00',
+              visitNotes: [
+                {
+                  type: 'VISIT_OUTCOMES',
+                  text: 'na',
+                },
+              ],
+              visitContact: {
+                name: 'Tess Bennett',
+                telephone: '0114 5555555',
+              },
+              visitors: [
+                {
+                  nomisPersonId: 4729570,
+                },
+                {
+                  nomisPersonId: 4729510,
+                },
+              ],
+              visitorSupport: [],
+              createdBy: 'user1',
+              createdTimestamp: '2022-05-20T15:29:04.997067',
+              modifiedTimestamp: '2022-05-20T15:51:49.983108',
+            },
+          ],
+        }
+
+        const social: Contact[] = [
+          {
+            personId: 4729510,
+            firstName: 'James',
+            lastName: 'Smith',
+            dateOfBirth: '1983-06-17',
+            relationshipCode: 'BRO',
+            relationshipDescription: 'Brother',
+            contactType: 'S',
+            contactTypeDescription: 'Social/ Family',
+            approvedVisitor: true,
+            emergencyContact: false,
+            nextOfKin: false,
+            addresses: [
+              {
+                addressType: 'Home Address',
+                street: 'Warren way',
+                town: 'Bootle',
+                postalCode: 'DN5 9SD',
+                country: 'England',
+                primary: true,
+                noFixedAddress: false,
+                startDate: '2021-03-01',
+                phones: [
+                  {
+                    number: '0113222333',
+                    type: 'BUS',
+                  },
+                  {
+                    number: '222333',
+                    type: 'HOME',
+                  },
+                ],
+                addressUsages: emptyUsages,
+              },
+            ],
+            restrictions: emptyRestrictions,
+          },
+          {
+            personId: 4729570,
+            firstName: 'Tess',
+            lastName: 'Bennett',
+            relationshipCode: 'AUNT',
+            relationshipDescription: 'Aunt',
+            contactType: 'S',
+            contactTypeDescription: 'Social/ Family',
+            approvedVisitor: true,
+            emergencyContact: false,
+            nextOfKin: false,
+            addresses: emptyAddresses,
+            restrictions: emptyRestrictions,
+          },
+        ]
+
+        orchestrationApiClient.getVisitsByDate.mockResolvedValue(pagedVisit)
+        prisonerContactRegistryApiClient.getPrisonerSocialContacts.mockResolvedValue(social)
+        const results = await visitService.getVisitsByDate({
+          username: 'user',
+          dateString: '2022-05-23',
+          prisonId,
+        })
+        const resultsCheck: {
+          extendedVisitsInfo: ExtendedVisitInformation[]
+          slots: {
+            openSlots: VisitsPageSlot[]
+            closedSlots: VisitsPageSlot[]
+            unknownSlots: VisitsPageSlot[]
+            firstSlotTime: string
+          }
+        } = {
+          extendedVisitsInfo: [
+            {
+              reference: 'ob-cw-lx-na',
+              prisonNumber: 'A8709DY',
+              prisonerName: '',
+              mainContact: 'UNKNOWN',
+              startTimestamp: '2022-05-23T09:00:00',
+              endTimestamp: '2022-05-23T09:29:00',
+              visitDate: '23 May 2022',
+              visitTime: '9am to 9:29am',
+              visitStatus: 'BOOKED',
+              visitRestriction: 'OPEN',
+              visitors: [
+                {
+                  personId: 4729510,
+                  name: 'James Smith',
+                  dateOfBirth: '1983-06-17',
+                  adult: true,
+                  relationshipDescription: 'Brother',
+                  address: 'Warren way,<br>Bootle,<br>DN5 9SD,<br>England',
+                  restrictions: [],
+                  banned: false,
+                },
+              ],
+            },
+            {
+              reference: 'lb-co-bn-oe',
+              prisonNumber: 'A8709DY',
+              prisonerName: '',
+              mainContact: 'Tess Bennett',
+              startTimestamp: '2022-05-23T10:00:00',
+              endTimestamp: '2022-05-23T11:00:00',
+              visitDate: '23 May 2022',
+              visitTime: '10am to 11am',
+              visitStatus: 'BOOKED',
+              visitRestriction: 'OPEN',
+              visitors: [
+                {
+                  personId: 4729510,
+                  name: 'James Smith',
+                  dateOfBirth: '1983-06-17',
+                  adult: true,
+                  relationshipDescription: 'Brother',
+                  address: 'Warren way,<br>Bootle,<br>DN5 9SD,<br>England',
+                  restrictions: [],
+                  banned: false,
+                },
+                {
+                  personId: 4729570,
+                  name: 'Tess Bennett',
+                  relationshipDescription: 'Aunt',
+                  address: 'Not entered',
+                  restrictions: [],
+                  banned: false,
+                },
+              ],
+            },
+          ],
+          slots: {
+            openSlots: [
+              {
+                visitTime: '9am to 9:29am',
+                visitType: 'OPEN',
+                sortField: '2022-05-23T09:00:00',
+                adults: 1,
+                children: 0,
+              },
+              {
+                visitTime: '10am to 11am',
+                visitType: 'OPEN',
+                sortField: '2022-05-23T10:00:00',
+                adults: 1,
+                children: 1,
+              },
+            ],
+            closedSlots: [],
+            unknownSlots: [],
+            firstSlotTime: '9am to 9:29am',
+          },
+        }
+
+        expect(orchestrationApiClient.getVisitsByDate).toHaveBeenCalledTimes(1)
+        expect(prisonerContactRegistryApiClient.getPrisonerSocialContacts).toHaveBeenCalledTimes(2)
+        expect(results).toEqual(resultsCheck)
       })
     })
   })
